cleans and organizes

+++ arch/arm/mach-s5pv210/adc.c
@@ -202,7 +202,7 @@ s3c_adc_read(struct file *file, char __user *buffer,
 }
 
 
-static int s3c_adc_ioctl(struct inode *inode, struct file *file,
+static long s3c_adc_ioctl(struct inode *inode, struct file *file,
 	unsigned int cmd, unsigned long arg)
 {
